Add user contact\_person field, update userid uniqueness validation, add Internal role, update permissions, and add can() function for checking user permissions in Vue.js
This commit adds a new field "contact\_person" to the

+++ routes/api.php
@@ -34,7 +34,7 @@ Route::post('register', [UserController::class, 'register']);
 
 Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::get('/get-permissions', function () {
-        return auth()->check()?auth()->user()->jsPermissions():0;
+        return auth()->check() ? auth()->user()->jsPermissions() : 0;
     });
 
 });
@@ -111,61 +111,63 @@ Route::group(['prefix' => 'supplier', 'middleware' => ['auth:sanctum', 'role:Sup
 Route::middleware(['auth:sanctum', 'role:Admin'])->group(function () {
 
     // Fetching Users
-    Route::get('users', [UserController::class, 'users']);
-    Route::get('supplier', [UserController::class, 'supplier']);
-    Route::get('buyer', [UserController::class, 'buyer']);
+    Route::get('users', [UserController::class, 'users'])->middleware('can:issueNewLoginIdPassword,setAccessAuthority,userManagement');
+    Route::get('supplier', [UserController::class, 'supplier'])->middleware('can:addNewSupplierEntry,setEditSupplierIDCode');
+    Route::get('buyer', [UserController::class, 'buyer'])->middleware('can:addNewBuyerEntry,editBuyerEntry');
     Route::get('supplierOrder', [UserController::class, 'supplierOrder']);
     Route::get('buyerOrder', [UserController::class, 'buyerOrder']);
     Route::get('supplier_profiles/{id}', [UserController::class, 'supplier_profiles']);
 
     // Adding Users
-    Route::post('addbuyers', [UserController::class, 'buyers']);
-    Route::post('addsupliers', [UserController::class, 'suppliers']);
+    Route::post('addbuyers', [UserController::class, 'buyers'])->middleware('can:addNewBuyerEntry,editBuyerEntry');
+    Route::post('addsupliers', [UserController::class, 'suppliers'])->middleware('can:setEditSupplierIDCode,editSupplierEntry');
 
     // Updating Users
-    Route::get('editusers/{id}', [UserController::class, 'usersEdit']);
-    Route::put('updateusers/{id}', [UserController::class, 'update']);
-    Route::post('addusers', [UserController::class, 'addUser']);
-    Route::delete('userDelete/{id}', [UserController::class, 'delUser']);
+    Route::get('editusers/{id}', [UserController::class, 'usersEdit'])->middleware('can:issueNewLoginIdPassword,setAccessAuthority,userManagement');
+    Route::put('updateusers/{id}', [UserController::class, 'update'])->middleware('can:issueNewLoginIdPassword,setAccessAuthority,userManagement');
+    Route::post('addusers', [UserController::class, 'addUser'])->middleware('can:issueNewLoginIdPassword,setAccessAuthority,userManagement');
+    Route::delete('userDelete/{id}', [UserController::class, 'delUser'])->middleware('can:issueNewLoginIdPassword,setAccessAuthority,userManagement');
 
     //Product-group
-    Route::post('product_group', [ProductController::class, 'product_group']);
-    Route::get('product_group_get', [ProductController::class, 'product_group_get']);
-    Route::get('product_group_get_all', [ProductController::class, 'product_group_get_all']);
+    Route::post('product_group', [ProductController::class, 'product_group'])->middleware('can:createProductGroup');
+    Route::get('product_group_get', [ProductController::class, 'product_group_get'])->middleware('can:createProductGroup');
+    Route::get('product_group_get_all', [ProductController::class, 'product_group_get_all'])->middleware('can:createProductGroup');
 
     //price_inquiry
-    Route::post('price_inquiry', [ProductController::class, 'price_inquiry']);
+    Route::post('price_inquiry', [ProductController::class, 'price_inquiry'])->middleware('can:createPriceInquiry');
     Route::post('update_price_inquiry/{id}', [ProductController::class, 'update_price_inquiry']);
     Route::get('price_inquiry_get', [ProductController::class, 'price_inquiry_get']);
     Route::delete('PriceDelete/{id}', [ProductController::class, 'PriceDelete']);
 
     //Orders
-    Route::post('orderentry', [OrderController::class, 'orderentry']);
+    Route::post('orderentry', [OrderController::class, 'orderentry'])->middleware('can:orderGeneralSinglePage,createNewOrder');
     Route::post('saveSelectedOrders', [OrderController::class, 'saveSelectedOrders']);
-    Route::get('orderentry', [OrderController::class, 'orderentryget']);
-    Route::get('orderAll', [OrderController::class, 'orderAll']);
-    Route::get('orderentry/{id}', [OrderController::class, 'orderentrygetID']);
-    Route::post('orderentry/{id}', [OrderController::class, 'orderUpdate']);
+    Route::get('orderentry', [OrderController::class, 'orderentryget'])->middleware('can:orderGeneralSinglePage,createNewOrder');
+    Route::get('orderAll', [OrderController::class, 'orderAll'])->middleware('can:miLungOrderPriceEnquiry');
+    Route::get('orderentry/{id}', [OrderController::class, 'orderentrygetID'])->middleware('can:editOrderDetails');
+    Route::post('orderentry/{id}', [OrderController::class, 'orderUpdate'])->middleware('can:editOrderDetails,voidOrder');
     Route::delete('orderDelete/{id}', [OrderController::class, 'orderDelete']);
     Route::post('placeAll', [OrderController::class, 'placeAll']);
     Route::get('shipmentsget', [ShipmentController::class, 'shipmentsget']);
     Route::get('Suppliers', [OrderController::class, 'Suppliers']);
-
+    Route::get('printview/{id}', [BuyerOrderController::class, 'printviewget'])->middleware('can:printviewConfirmRejectButton');
+    Route::post('printview/{id}', [BuyerOrderController::class, 'printview'])->middleware('can:printviewConfirmRejectButton');
+    Route::post('masscargo/{id}', [SupplierOrderController::class, 'masscargo'])->middleware('can:massCargoPhotoApproval');
     //Shipment
-    Route::post('create_so', [ShipmentController::class, 'create_so']);
-    Route::delete('soDelete/{id}', [ShipmentController::class, 'soDelete']);
-    Route::get('shipmentget', [ShipmentController::class, 'shipmentget']);
-    Route::get('shipmentget/{id}', [ShipmentController::class, 'shipmentgetid']);
-    Route::post('shipment/{id}', [ShipmentController::class, 'shipment']);
+    Route::post('create_so', [ShipmentController::class, 'create_so'])->middleware('can:createSONumber');
+    Route::delete('soDelete/{id}', [ShipmentController::class, 'soDelete'])->middleware('can:editShipmentOverview');
+    Route::get('shipmentget', [ShipmentController::class, 'shipmentget'])->middleware('can:shipmentOverview,editShipmentOverview');
+    Route::get('shipmentget/{id}', [ShipmentController::class, 'shipmentgetid'])->middleware('can:editShipmentOverview');
+    Route::post('shipment/{id}', [ShipmentController::class, 'shipment'])->middleware('can:editShipmentOverview');
     // Information
-    Route::post('create_doc', [ShipmentController::class, 'create_doc']);
-    Route::get('information/{id}', [ShipmentController::class, 'information']);
+    Route::post('create_doc', [ShipmentController::class, 'create_doc'])->middleware('can:exportShippingDocuments');
+    Route::get('information/{id}', [ShipmentController::class, 'information'])->middleware('can:exportShippingDocuments');
 
     // Finance
     Route::get('so/{id}', [ShipmentController::class, 'SupplierSo']);
-    Route::get('invoice/{id}', [ShipmentController::class, 'invoice']);
-    Route::get('buyerFinance', [UserController::class, 'buyerFinance']);
-    Route::post('rcvablesave/{id}', [ShipmentController::class, 'rcvablesave']);
+    Route::get('invoice/{id}', [ShipmentController::class, 'invoice'])->middleware('can:accountReceivable');
+    Route::get('buyerFinance', [UserController::class, 'buyerFinance'])->middleware('can:accountReceivable');
+    Route::post('rcvablesave/{id}', [ShipmentController::class, 'rcvablesave'])->middleware('can:accountReceivable');
 
     // Statistics
     Route::get('statfilter', [StatController::class, 'statfilter']);
@@ -178,9 +180,9 @@ Route::middleware(['auth:sanctum', 'role:Admin'])->group(function () {
     Route::post('statement', [ProductController::class, 'statement']);
 
     // Adding Products
-    Route::get('products', [UserController::class, 'products']);
-    Route::delete('prodDelete/{id}', [ProductController::class, 'prodDelete']);
-    Route::post('addprod', [ProductController::class, 'addprod']);
+    Route::get('products', [UserController::class, 'products'])->middleware('can:addProductEntry,editProductEntry,accessImportExportCertificateTestingReport');
+    Route::delete('prodDelete/{id}', [ProductController::class, 'prodDelete'])->middleware('can:addProductEntry,editProductEntry,accessImportExportCertificateTestingReport');
+    Route::post('addprod', [ProductController::class, 'addprod'])->middleware('can:addProductEntry,editProductEntry,accessImportExportCertificateTestingReport');
 
 });
